+++ app/controllers/project_information_controller.rb
@@ -1,5 +1,5 @@
 class ProjectInformationController < ApplicationController
   def show
-    @project = Project.find(params[:project_id])
+    @project = ProjectPresenter.new(Project.find(params[:project_id]))
   end
 end