+++ app/models/project.rb
@@ -9,6 +9,8 @@ class Project < ApplicationRecord
   validates :incoming_trust_ukprn, presence: true, numericality: {only_integer: true}
   validates :target_completion_date, presence: true
   validates :incoming_trust_ukprn, ukprn: true
+  validates :advisory_board_date, presence: true, on: :create
+  validates :advisory_board_date, past_date: true
 
   validate :first_day_of_month
   validate :target_completion_date_is_in_the_future, on: :create