+++ app/validators/past_date_validator.rb
@@ -1,0 +1,9 @@
+class PastDateValidator < ActiveModel::EachValidator
+  def validate_each(record, attribute, value)
+    return unless value.present?
+
+    unless value.past?
+      record.errors.add(attribute, :must_be_in_the_past)
+    end
+  end
+end