+++ spec/models/project_spec.rb
@@ -8,6 +8,8 @@ RSpec.describe Project, type: :model do
     it { is_expected.to have_db_column(:caseworker_id).of_type :uuid }
     it { is_expected.to have_db_column(:team_leader_id).of_type :uuid }
     it { is_expected.to have_db_column(:caseworker_assigned_at).of_type :datetime }
+    it { is_expected.to have_db_column(:advisory_board_date).of_type :date }
+    it { is_expected.to have_db_column(:advisory_board_conditions).of_type :text }
   end
 
   describe "Relationships" do
@@ -70,6 +72,8 @@ RSpec.describe Project, type: :model do
 
   describe "Validations" do
     before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
+
+    it { is_expected.to validate_presence_of(:advisory_board_date).on(:create) }
 
     describe "#urn" do
       it { is_expected.to validate_presence_of(:urn) }