+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- The Handover/new project form collects the advisory board date and any
+  conditions.
+
+### Changed
+
+### Fixed
+
 ## [Release 3][release-3]
 
 ### Added